+++ kotlin/reakt-native-toolkit/src/iosMain/kotlin/de/voize/reaktnativetoolkit/util/PlatformUtil.kt
@@ -1,0 +1,7 @@
+package de.voize.reaktnativetoolkit.util
+
+import platform.Foundation.NSUUID
+
+internal actual fun uuid(): String {
+    return NSUUID().UUIDString().lowercase()
+}